Address feedback from @benjamn regarding setRunningTest.
A great suggestion!

Refs: https://github.com/meteor/meteor/pull/9364#discussion_r150981769
